Keep the listing state type inside the membership module
Exporting it widened the surface without a reader, and the field
policy gate held its unused cap field up instead. The package writes
sibling infers the states from the readers' returns, so the type stays
module-private.
