+++ src/features/admin/api-groups.ts
@@ -7,7 +7,7 @@ import {
   type PackageFlags,
   readPackageFlagsTxOrNull,
   writePackageMembersTx,
-} from "#db/groups/membership.ts";
+} from "#db/groups/membership/package-writes.ts";
 import {
   generateUniqueGroupSlug,
   groups,
