+++ test/features/admin/listings-parents.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "@std/expect";
 import { it as test } from "@std/testing/bdd";
 import { executeWithoutCacheInvalidation } from "#db/client.ts";
-import { assignListingsToGroup } from "#db/groups/membership.ts";
+import { assignListingsToGroup } from "#db/groups/membership/package-writes.ts";
 import { listingChildren, listingParents } from "#db/listing-parents.ts";
 import { t } from "#i18n";
 import {
