+++ test/integration/routes/read-only.test.ts
@@ -177,7 +177,7 @@ describeWithEnv(
     test("groups on listings page show Registration Closed in read-only mode", async () => {
       const { groups } = await import("#db/groups.ts");
       const { assignListingsToGroup } = await import(
-        "#db/groups/membership.ts"
+        "#db/groups/membership/package-writes.ts"
       );
       const { listingsTable } = await import("#db/listings/records.ts");
       const { hmacHash } = await import("#crypto/hashing.ts");
