+++ test/integration/server/ledger/filters-stats.test.ts
@@ -3,7 +3,7 @@ import { it as test } from "@std/testing/bdd";
 import { KIND } from "#accounting/kinds.ts";
 import { MANUAL_LISTING_COST } from "#accounting/manual-entries.ts";
 import { postTransferGroups } from "#accounting/store.ts";
-import { assignListingsToGroup } from "#db/groups/membership.ts";
+import { assignListingsToGroup } from "#db/groups/membership/package-writes.ts";
 import { account } from "#shared/ledger/account.ts";
 import {
   listingMoneyLegs,
