+++ test/shared/db/groups/set-groups.test.ts
@@ -7,7 +7,7 @@
 import { expect } from "@std/expect";
 import { it as test } from "@std/testing/bdd";
 import { withTransaction, writeRowInTransaction } from "#db/client.ts";
-import { assignListingsToGroup } from "#db/groups/membership.ts";
+import { assignListingsToGroup } from "#db/groups/membership/package-writes.ts";
 import {
   getGroupPackagePrices,
   getListingsByGroupId,
